Move constraints into pepper2.common
There is no daemon-specific code, so moved to common
