+++ pepper2/daemon/udisks_manager.py
@@ -10,9 +10,9 @@ from typing import Dict, List, Type
 
 from pydbus.bus import Bus
 
+from pepper2.common.constraint import Constraint
 from pepper2.common.drive_types import DRIVE_TYPES
 from pepper2.common.drives import Drive, DriveType
-from pepper2.daemon.constraint import Constraint
 from pepper2.daemon.dbus.controller import Controller
 
 LOGGER = logging.getLogger(__name__)
